src/Makefile: call clean in correct uimac directory src/uimacbridgenew.ml: correct a comment src/Makefile.Ocaml: don't use sed anymore to create the Info.plist, it's done by the XCode project (thus adding the Info.plist file needed) src/uimacnew/uimacnew.xcodeproj: the new XCode project, cleaned up src/uimacnew/*.m: patches by Gregory Entin unison-2.27.57-root-prop.diff:
Fixes the crash with objectAtIndex: 0 out of bounds error, that seems
to be recently reported to this list. I can reproduce it by modifying
permissions of two otherwise identical *root* folders (e.g. I use g+w
for one folder and g-w for another). The thing is "pathArray" is empty
in that case (see the patch).

unison-2.27.57-uimacnew-utf8.diff:

Enforces UTF8 instead of "default" C-string encoding for some strings
in Mac UI (seems to be crucial to deal with non-ASCII file names under
Mac OS X).

unison-2.27.57-symroot.diff:

Fixes xcodebuild invocation in case when Xcode is configured with
non-default location for build directories.

Thanks to Steve Kalkwarf for the help with XCode



git-svn-id: https://webdav.seas.upenn.edu/svn/unison@276 4bce34ff-96ee-0310-b826-95abcfff6e26